Move drawing functions to 16bit colour - better theme colours (An_dz, from Standard)
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@8134 8aca7d54-2c30-db11-9de9-000461428c89

+++ simfab.cc
@@ -2956,7 +2956,7 @@ void fabrik_t::new_month()
 }
 
 // static !
-unsigned fabrik_t::status_to_color[MAX_FAB_STATUS] = {
+uint8 fabrik_t::status_to_color[MAX_FAB_STATUS] = {
 	COL_WHITE, COL_GREEN, COL_DODGER_BLUE, COL_LIGHT_TURQUOISE, COL_BLUE, COL_DARK_GREEN,
 	COL_GREY3, COL_DARK_BROWN+1, COL_YELLOW-1, COL_YELLOW, COL_ORANGE, COL_ORANGE_RED, COL_RED, COL_STAFF_SHORTAGE };
